requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
anthropic==0.62.0
python-dotenv==1.0.0
httpx==0.25.0
torch==2.8.0
torchvision==0.23.0
scikit-learn==1.6.1
numpy==2.0.2